Appointment Page and Calendar is added to enable users to get appointments.
Search Query for psychologists is changed.
A bug in the seeder causing same patients to be inserted into psychologist object is fixed.
A bug at Login page for verified psychologists is fixed.
Profile photos are now fetched from the database.
Appointment Creation page is added for Client.
Calendar to pick up appointments is implemented.
Registered Therapist List is now available to get appointments; however, not functional.